Move StorageCollection to storage module
