+++ redfish/src/computer_system/storage.rs
@@ -24,6 +24,9 @@ use crate::ResourceSchema;
 use nv_redfish_core::Bmc;
 use nv_redfish_core::NavProperty;
 use std::sync::Arc;
+use crate::schema::storage_collection::StorageCollection as StorageCollectionSchema;
+use crate::patch_support::CollectionWithPatch;
+use crate::schema::resource::ResourceCollection;
 
 /// Represents a storage controller in a computer system.
 ///
@@ -83,5 +86,55 @@ impl<B: Bmc> Storage<B> {
 impl<B: Bmc> Resource for Storage<B> {
     fn resource_ref(&self) -> &ResourceSchema {
         &self.data.as_ref().base
+    }
+}
+
+/// Storage collection.
+///
+/// Provides functions to access collection members.
+#[cfg(feature = "storages")]
+pub struct StorageCollection<B: Bmc> {
+    bmc: NvBmc<B>,
+    collection: Arc<StorageCollectionSchema>,
+}
+
+#[cfg(feature = "storages")]
+impl <B:Bmc> StorageCollection <B> {
+    
+    /// Create a new storage collection handle.
+    pub(crate) async fn new(
+        bmc: &NvBmc<B>,
+        nav: &NavProperty<StorageCollectionSchema>,
+    ) -> Result<Self, Error<B>> {
+        let collection = Self::expand_collection(bmc, nav, None, None).await?;
+        Ok(Self {
+            bmc: bmc.clone(),
+            collection,
+        })
+    }
+
+    /// List all storages available in this BMC.
+    ///
+    /// # Errors
+    ///
+    /// Returns an error if fetching storage data fails.
+    pub async fn members(&self) -> Result<Vec<Storage<B>>, Error<B>> {
+        let mut members = Vec::new();
+        for m in &self.collection.members {
+            members.push(Storage::new(&self.bmc, m).await?);
+        }
+        Ok(members)
+    }
+}
+
+#[cfg(feature = "storages")]
+impl<B: Bmc> CollectionWithPatch<StorageCollectionSchema, StorageSchema, B>
+    for StorageCollection<B>
+{
+    fn convert_patched(
+        base: ResourceCollection,
+        members: Vec<NavProperty<StorageSchema>>,
+    ) -> StorageCollectionSchema {
+        StorageCollectionSchema { base, members }
     }
 }
